backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "paho-mqtt>=2.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
